requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9

streamlit==1.38.0
pandas==2.2.2
matplotlib==3.9.2

openai==1.47.0
google-generativeai==0.7.2
python-dateutil==2.9.0.post0
